Rely on FindHDF5 rather than h5fc and h5pfc scripts
The FindHDF5.cmake packaged with CMake is broken in a number of respects. The HL
components don't work (debian has a patch in cmake-data). It's also impossible
to prefer a parallel installation if both h5fc and h5pfc appear on your
PATH. Finally, hdf5_hl is not included in the list of libraries needed for the
Fortran_HL component. A local version of FindHDF5 is used here which fixes all
these issues. This enables one to compile OpenMC with MPI + serial HDF5 if
needed by introducing the PHDF5 preprocessor flag.