Add 'coincident' argument to distance routines.
When a particle is advanced to a surface, due to limited precision it may appear
as though the particle's position is not exactly coincident (as tested by the
sense of the particle with respect to the surface). Thus, the only sure way of
making sure that distance() returns the correct value is to pass it an argument
indicating whether the particle is coincident with the surface.