Improve check for coincidence in sense().
The previous solution was to move the particle forward artificially and then
call sense() recursively. A better way is to look at the projection of the
particle's direction on the surface normal from which the sign determines which
side the particle will be on when it moves forward. This enables us to make
sense() a pure non-recursive function.