Don't pre-compute kappa-fission cross sections.
This also fixes a bug in the kappa-fission score. Before, kappa-fission was
computed as Q*fission, but this was done before URR cross sections were
determined. Thus, if fission changed in calculate_urr_xs, this wasn't reflected
in the kappa-fission score. Now, since it is all done at tally-time, there is no
inconsistency.