Add basic windowed multipole library support
This commit adds the ability to load in windowed multipole
temperature-dependent cross section data.  Given a windowed multipole
file named ZAID.h5 in the same directory as the *.xml files of a
simulation, OpenMC will now load this HDF5 file.  This will replace the
resolved resonance region with windowed multipole data, and all other
pointwise data by a small pointwise library to minimize memory
footprint.

Temperatures are tenatively set by adding a <temperature> tag inside of
a material in materials.xml.  Units are in Kelvin for now.

This commit also adds the Faddeeva.cc library from
http://ab-initio.mit.edu/wiki/index.php/Faddeeva_Package which requires
OpenMC to compile both Fortran and C code.  Modifications have been done
to most of CMakeLists.txt, but not all of it.