Fix tracklength mesh tallies
Unfortunately this required a bit of a hack. Since the filters do not know the
length of the tack, they need to get the old coordinates from p % last_xyz. But
p % last_xyz isn't updated at every event. Changing the tracking so p % last_xyz
is updated at every event results in CMFD SIGFPEs, so we need to keep the old
p % last_xyz as p % last_xyz_current