Add test for the instantiation of deplete.Operator
Construct a minimal cross sections xml file that only
contains the depletion_chain node required by
the Operator. The path to this file is set to the
OPENMC_CROSS_SECTIONS file, and reverted after the test.

The depletion_chain points towards the chain_simple.xml
files, and a reference Chain is produced in the test.
The test involves creating a bare Operator instance, that
constructs a chain based on the temporary
OPENMC_CROSS_SECTIONS file. This chain is compared to
that produced by reading the chain_simple.xml file stored
in the test directory.