Support depletion_chain in DataLibrary to/from xml and register
DataLibrary.from_xml looks for depletion_chain node in the
cross sections xml file. If found the path to this file is added
to the list of libraries. An empty list of materials is used
for consistency with other libraries. Type and path attributes
are consistent.

DataLibrary.export_to_xml does not write the materials
attribute if the value stored in that library evaluates to False,
e.g. for the empty list stored in the depletion chain.

DataLibrary.register_file has a conditional wherein a passed xml
file is treated as the depletion_chain and hdf5 files are
treated as material/isotopic cross section files.