Iterate over all libraries, not children, in Element.expand
After PR #1247, the cross section file is allowed to contain
the depletion chain file. This is present as a new tag, of
the form
```
   <depletion_chain path="path.xml" type="depletion_chain" />
```

The Element.expand method looked for `material` attributes
in all children of the OPENMC_CROSS_SECTION file.
This caused an error for the depletion_chain element.
Now, only those children with the `library` tag are
iterated over.