Update test_deplete_predictor to check depletion time
Due to where the results are saved for the predictor routine,
at the top of the for loop vs. the end, the positions
are incremented by one in the depletion time. This causes
a value of zero to be returned as the first item in the array
from get_depletion_time.

This save was moved in commit 87aed7a0b095b3902156e2a385bc46db2ba7f5ac
of PR #1014 to make the restart feature easier.