Add predictor-specific logic in get_depletion_time
The predictor saves BOS number densities, eigenvalues, and
other results before the depletion event, unlike all
other schemes, which have at least one depletion event between
the BOS transport solution and the depletion solve.
This causes the first value in the resulting depletion process
time vector to be zero, and the last value to be non-zero,
opposite to other schemes.