Add a litle tolerance for capture branching ratios
When adding capture branching ratios to a ``Chain``,
the sum of ratios for a given parent nuclide was allowed to be no
greater than 1.0. This limit is very tight, and does not allow
for minor floating point precision differences, like
1.0000000002.

The maximum value the sum of branching ratios from a single
parent is now 1.00001, with no equality check. This has
been added to the method docstring for transparency.

Some content was removed from the docstring that implied that
``branch_ratios`` would be modified in place. This is not the case,
and thus the content was removed.