Pass iteration index to Integrator.__call__
For the higher order methods, like LEQI, a lower order
method may be called for the first few steps. The LEQI
method, with no restart data, will call CELI for the first
iteration. To aid this, concrete Integrator classes will be
passed an integer for the current iteration.

Some methods, like the CECM and CELI, do not need this information.
To maintain a consistent API, the iteration is an optional value
to be passed to CECM.__call__, but will not be used. The default
value in the call signature is ``_i=-1``, with documentation repeating
that the value is not used nor needed.