Use correct material index in operator unpacking
Fix a bug introduced where the material iteration index,

        for i, mat in enumerate(self.local_mats):

was used for the global __material__ index. The value of i
was used to extract reaction rate tallies for that material.
This causes an issue with multiple operators on MPI processes,
where an Operator`s material i does not equal the i-th burnable
material.