Polish up energy deposition setting: export, tests
Changed the name of the setting from fiss_edep_mode to
energy_deposition in the c and python sources. Default option
is not local photons, not local photon. Add exporting to
and reading from xml methods related to this setting.

Include this setting in tests/unit_tests/test_settings.py
and settings.xml source format