Provide Integrator get bos data from openmc|restart methods
All of the integrators pull the previous concentration and
operator results from a restart file in the same way, but vary
in how subsequent steps are handled. The predictor saves data
here, while the SIE-based methods scale and reset the particles
used. To this end, the _get_bos_data function has been broken into
two methods: one for calling the operator and one for pulling from
a restart file.