Return BOS concentration as list of vectors from restarts
Operator.initial_condition return list of vectors representing
the initial composition for a burnable material as an element
in the list. This format is expected from depletion solvers as well.
Previously, when pulling from the previous results on an operator,
the initial concentration was returned simply as a 2D array of
compositions. This commit converts the 2D array to a list of
vectors to be consistent across the integration schemes.