Avert bad numpy slices unpacking fission yields with no materials
For the case where there are more processes than burnable materials,
e.g. single pin cell or assembly with one burnable material, then
the local_indexes attribute on some FissionYieldHelpers will
be an empty array. This causes errors trying to use this as an
indexer into the tally results.

Instead, simply return from the unpacking if this is the case.
The compute_yields function is called inside the Operator's
unpacking of it's local materials, so that __should__ not be
in a position to call compute_yields for a process with no
burnable materials.