Determine fission yield helper based on Operator arguments
A single argument, fission_yield_mode, is used to determine
what type of FissionYield to use on the Operator. The options
are

* "constant": ConstantFissionYieldHelper [default]
* "cutoff": FissionYieldCutoffHelper

An additional argument, fission_yield_opts, can be supplied
as a dictionary of keyword arguments to pass to the new
helper. This allows the user to have control over how
the private helper is created.