Leave He4 alone in Chain.set_branch_ratios
In an (n, alpha) reaction, the branching ratio for helium should
always be one, as it is produced for every reaction.
It is the branching ratio of the other targets that must be modified.

He4 is treated as a secondary particle for this case, and its
reaction is not modified nor removed from the parent.