Use helper function when scoring fission_q_*
Add a brief helper function get_nuc_fission_q in
src/tallies/tally_scoring.cpp that is responsible for
returning the correct fission q value for a given
score. The two allowed scores are SCORE_FISS_Q_PROMPT and
SCORE_FISS_Q_RECOV. The helper function examines the correct
dataset on the nuclide object and evaluates the function given
the incident particle energy or returns zero.
The default value of zero is returned if the fission_q_prompt_ or
fission_q_recov_ datasets are empty, or the score is not
one of the expected types.