Add openmc::Tally.writeable_ attribute
Boolean flag that indicates if a tally should be written to
output files. This has no effect on tally scoring, but instead
will be used to skip writing Tally data for those not marked
as writeable.

Related to #1327 as the depletion interface will use this flag
to reduce time spent writing reaction rate tallies and also
reducing the size of the final statepoint files