Teach openmc_statepoint_write about internal tallies
If a tally is marked as not writeable, a tally group
will still be added to the statepoint file. However,
no filter, score, nuclide, or results data will be written to
this tally group. A new attribute is introduced to this tally
group "internal" that is an integer flag: 0 if the tally
is not internal (i.e. is writeable) and 1 if the tally is
internal (i.e. is not writeable).

The internal attribute has been added to the statepoint
io format in the documentation.

Other changes: use
```
const auto& tally : model::tallies

  tally->X
```
to iterate over tallies during the writing process