Teach load_state_point about internal tallies
When reading in tally data from a statepoint file, a check
is performed on the "internal" attribute of the tally group.
If the attribute doesn't exist (previous statepoint files), then
it is assumed that the tally is writeable / not internal. Otherwise,
the internal attribute is read from the dataset.

If the tally is determined to be internal, the value of
tally.writeable_ is set to false and no tally data
is read